Updated game map object for faster sprite loading.
The gamemap object previously 'reinitialized' all sprites (recreated
them) with every press. Now, instead, the game keeps track of the
player's FOV and checks every tick to see if the FOV has changed. If it
has, it proceeds to load new sprites in view and delete old sprites out
of view. As a result, things run *much* faster.