Re-added the turn queue (entities now take turns)
The turn queue (located in event.py) is up and running now; the list is
sorted (using a GeoNode's time attribute--higher time values go first)
and if the first entry is *not* the current stack focus (IE, player's
character), the entity in question takes its turn (firing an EntityTurn
event), which eats up some of its time. The list is then sorted
again--this continues until the current turn is the stack's focus (IE,
player), at which point the turn queue just shuts up and waits for the
player to take its turn.

The cool thing about this means there is no fundamental difference
between player turns and monster turns except that the turn queue shuts
up and listens when it's the player's turn.