Added command delay.
Added a delay for user commands on the game map (to prevent a glut of
commands choking the UI). Every time you give a command, it goes into a
list; the list fires the first command on it whenever the focus node's
sprite isn't moving. Also cleaned up/changed a little code here and
there in the gamemap--fixed a glitch with the Menu window (it was
getting drawn over by the gamemap)--and rewrote the gamemap's refresh
method a little.