Basic inventory, throwing items.
So, I added an inventory window (which comes complete with a random set
of potions) and set up the cursor to allow you to click on an inventory
item (thereby placing it 'in' the cursor), click elsewhere on the
inventory (thereby 'dropping' it back into the inventory), or click
elsewhere on the map (thereby 'throwing' it at that particular tile). I
haven't gotten the cursor to display the item you're currently
manipulating yet, though, so when your cursor is 'holding' an object, it
looks exactly like it does when it's not holding an object.

I'm probably also going to want to make this whole throwing thing part
of the command delay chain. Or disable throwing until your sprite has
stopped moving.