Fixes to turn_queue and stack.
I shifted the sort() method for the turn_queue (which determines who's
turn it is) into the SpendTime event (which means we only check to see
who's turn it is when someone just took their turn). Also shifted the
command delay (in the gamemap object) to start firing commands whenever
it's the player's turn--and to continue firing them until there are no
commands or it's no longer the player's turn (this way, I know the
player's actions will always happen in appropriate order). Set up an
image library in the stack object for faster loading (it loads and
scales the images based on screen size--and 'reloads' and 'rescales'
them in case of a re-initialization).

Noticed a glitch--got a key error in my entities graph (in the area
object). This implies an entity wasn't in the position it was supposed
to be, which is *very* bad, but I only got it once and haven't been able
to repeat it. Will keep an eye out for it, though.