Fixed another command glitch
This one was silly (I need to start testing more extensively before I do
my commits/pushes); I forgot to double-check to make sure the command
list was not empty, meaning on the user's turn, we kept firing the first
command in the list even if there was no more commands (resulting in a
loop). In other words, if you gave an invalid command (walked into a
wall), the game would lock up.

Fixed now.