Added death, float text layering
Made it so entities now have a die() method to remove them from play on
death; also added CheckForDeath and EntityDeath as events that carefully
remove entities from play (producing floating text and removing sprites
that are on screen).

Changed the children attribute of the gamemap to a list instead of a set
after I realized that the order of floating text is important (you want
the most 'current' floating text on top).