Added Sheet, changed stats.
So, I decided I didn't like the way I was handling strength, strength
modifiers, etc--so I added all of this stuff under the sheet.py module,
which has 'Abilities', 'Modifiers', 'Defenses', and 'Attacks'
respectively--taking care of the upkeep for all of these things. I might
muck around with this later.