Fixed 'death' glitch.
'Dead' text wasn't popping up because I was deleting the sprite before
it had a chance to fire the appropriate object on its path. Just took
out the deletion and replaced it with a command that adds the 'destroy
self' command to the sprite's path (so it'll execute it last).